Added tests for the custom api token available to administrator
